Remove port of Montgomery conversion from Adam's code
This should be brought back later as part of reworking the compressed
point formats / serialization code.  Right now there's just
"CompressedPoint" which is in the ed25519 format.  Ideally, users should
be able to serialize points to formats used for X25519, for decaf, etc.
and not have to worry too much about the internal model.