Add iterator length checks to multiscalar muls.
This partially re-adds functionality removed in commit

1b205c27f56f3b3e3766fe36c88e8d8ff012fdcd

We would like to require ExactSizeIterator, but unfortunately we can't
do that, since ExactSizeIterators aren't chainable, for (in my opinion)
silly reasons (chaining two 4-billion-element ExactSizeIterators could
overflow on 32-bit systems).  Instead we inspect the size hints manually
and assert that the lower and upper bounds are all equal.